//Mahesh

+++ src/main/java/com/ats/webapi/model/RegularSpecialStockCal.java
@@ -14,24 +14,24 @@ public class RegularSpecialStockCal {
 	
 	@Id
 	@Column(name="reg")
-	private int reg;
+	private float reg;
 	
 	@Column(name="sp")
-	private int sp;
+	private float sp;
 
-	public int getReg() {
+	public float getReg() {
 		return reg;
 	}
 
-	public void setReg(int reg) {
+	public void setReg(float reg) {
 		this.reg = reg;
 	}
 
-	public int getSp() {
+	public float getSp() {
 		return sp;
 	}
 
-	public void setSp(int sp) {
+	public void setSp(float sp) {
 		this.sp = sp;
 	}
